refactor(account-bootstrap): own bootstrap model types
