+++ src/services/accounts/autoDetectCompletion/completion.ts
@@ -2,10 +2,10 @@ import {
   AUTO_DETECT_FAILURE_REASONS,
   type AutoDetectFailureReason,
 } from "~/constants/autoDetect"
+import type { SiteStatusInfo } from "~/services/accountBootstrap/model"
 import { getSiteName } from "~/services/accounts/siteName"
 import type { AccountCompletionHelpers } from "~/services/apiAdapters/contracts/accountCompletion"
 import { getSiteAdapter } from "~/services/apiAdapters/registry"
-import type { SiteStatusInfo } from "~/services/apiService/common/type"
 import { API_SERVICE_FETCH_CONTEXT_KINDS } from "~/services/apiTransport/type"
 import type {
   ApiServiceFetchContext,
