+++ src/services/apiService/aihubmix/index.ts
@@ -1,6 +1,11 @@
 import { AIHUBMIX_API_ORIGIN, SITE_TYPES } from "~/constants/siteType"
 import { UI_CONSTANTS } from "~/constants/ui"
 import type {
+  AccessTokenInfo,
+  SiteStatusInfo,
+  UserInfo,
+} from "~/services/accountBootstrap/model"
+import type {
   AccountData,
   RefreshAccountResult,
   TodayIncomeData,
@@ -15,12 +20,7 @@ import {
 } from "~/services/accountTokens/apiTokenKey"
 import { determineHealthStatus } from "~/services/apiService/common"
 import { API_ERROR_CODES, ApiError } from "~/services/apiService/common/errors"
-import type {
-  AccessTokenInfo,
-  ApiServiceAccountRequest,
-  SiteStatusInfo,
-  UserInfo,
-} from "~/services/apiService/common/type"
+import type { ApiServiceAccountRequest } from "~/services/apiService/common/type"
 import { fetchApiData } from "~/services/apiService/common/utils"
 import type {
   ApiResponse,
